backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
# For serving static files in dev/prod
whitenoise==6.7.0
